requirements.txt
Flask
Flask-Login
pyodbc
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
plotly
joblib
Werkzeug
python-dotenv
bcrypt
google-generativeai 
requests



